+++ robot-core/src/main/java/org/obolibrary/robot/ExportOperation.java
@@ -1,7 +1,6 @@
 package org.obolibrary.robot;
 
 import com.google.common.collect.Lists;
-import java.io.*;
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -945,8 +944,9 @@ public class ExportOperation {
       switch (colName) {
         case "IRI":
           String iriString = entity.getIRI().toString();
-          if (format.equalsIgnoreCase("html")) {
-            String display = String.format("<a href=\"%s'\">%s</a>", iriString, iriString);
+          if (format.toLowerCase().startsWith("html")) {
+            String display =
+                String.format("<a href=\"%s\">%s</a>", iriString, iriString.replace("&", "&amp;"));
             cell = new Cell(col, display, iriString);
           } else {
             cell = new Cell(col, iriString);